app/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0